+++ components/Header/SearchBar.tsx
@@ -86,7 +86,7 @@ const SearchBar: React.FC<{ className?: string }> = ({ className = "" }) => {
                     <ul className="absolute top-full mt-2 w-full bg-white dark:bg-dark shadow-lg rounded-lg z-10 shadow-slate-300 dark:shadow-slate-800">
                         {results.map((result) => (
                             <li key={result.id} className="flex items-center p-2 border-b border-slate-200 dark:border-slate-700 hover:bg-slate-200 dark:hover:bg-slate-700">
-                                <Link href={`/blogs/${result.slug}`} className="flex items-center w-full" onMouseDown={(e) => e.preventDefault()} onClick={handleResultClick}>
+                                <Link href={`/blog/${result.slug}`} className="flex items-center w-full" onMouseDown={(e) => e.preventDefault()} onClick={handleResultClick}>
                                     <span className="text-slate-800 dark:text-light">{result.title}</span>
                                 </Link>
                             </li>
@@ -114,7 +114,7 @@ const SearchBar: React.FC<{ className?: string }> = ({ className = "" }) => {
                             <ul className="bg-white dark:bg-dark shadow-lg rounded-lg z-10 shadow-slate-300 dark:shadow-slate-800">
                                 {results.map((result) => (
                                     <li key={result.id} className="flex items-center p-2 border-b border-slate-200 dark:border-slate-700 hover:bg-slate-200 dark:hover:bg-slate-700">
-                                        <Link href={`/blogs/${result.slug}`} className="flex items-center w-full" onMouseDown={(e) => e.preventDefault()} onClick={handleResultClick}>
+                                        <Link href={`/blog/${result.slug}`} className="flex items-center w-full" onMouseDown={(e) => e.preventDefault()} onClick={handleResultClick}>
                                             <span className="text-slate-800 dark:text-light">{result.title}</span>
                                         </Link>
                                     </li>